refactor: split EditPanel.svelte into composable sections
Decompose the 1,754-line EditPanel.svelte into five black-box section
components (EditPanelRack, EditPanelMetadata, EditPanelPosition,
EditPanelImage, EditPanelActions); the host shrinks to a selection
router. Sections are single-entity and compose into the future M14
Edit tab (#2077). Pure structural refactor, no behaviour change.

- Add shared SelectedDeviceInfo type to src/lib/types
- Sections access stores via singleton getters; only resolved
  selection is passed as props
- Document the section contract in the M14 spec
- EditPanelRack still hosts the transitional rear-view + annotation
  toggles (the View tab #2078 absorbs them)

Closes #1398
